fix(cli): isolate session directory fallback
