+++ packages/opencode/src/session/index.ts
@@ -9,7 +9,7 @@ import { Config } from "../config/config"
 import { Flag } from "../flag/flag"
 import { Installation } from "../installation"
 
-import { Database, NotFoundError, eq, and, gte, isNull, desc, like, or } from "../storage/db" // kilocode_change
+import { Database, NotFoundError, eq, and, gte, isNull, desc, like } from "../storage/db"
 import { SyncEvent } from "../sync"
 import { SessionTable } from "./session.sql"
 import { Storage } from "@/storage/storage"
@@ -25,7 +25,6 @@ import { Snapshot } from "@/snapshot"
 import { ProjectID } from "../project/schema"
 import { WorkspaceID } from "../control-plane/schema"
 import { SessionID, MessageID, PartID } from "./schema"
-import { Filesystem } from "../util/filesystem" // kilocode_change: normalize directory for Windows drive-letter casing
 import { KiloSession } from "@/kilocode/session" // kilocode_change
 
 import type { Provider } from "@/provider/provider"
@@ -795,23 +794,11 @@ export namespace Session {
     limit?: number
   }) {
     const project = Instance.project
-    // kilocode_change start: preserve directory lookups across project id changes
-    const dir = input?.directory ? Filesystem.resolve(input.directory) : undefined
-    const conditions = [
-      dir
-        ? or(eq(SessionTable.project_id, project.id), eq(SessionTable.directory, dir))
-        : eq(SessionTable.project_id, project.id),
-    ]
-    // kilocode_change end
+    const conditions = KiloSession.filters({ projectID: project.id, directory: input?.directory }) // kilocode_change
 
-    // kilocode_change start: preserve directory lookups across project id changes
     if (input?.workspaceID) {
       conditions.push(eq(SessionTable.workspace_id, input.workspaceID))
     }
-    if (dir) {
-      conditions.push(eq(SessionTable.directory, dir))
-    }
-    // kilocode_change end
     if (input?.roots) {
       conditions.push(isNull(SessionTable.parent_id))
     }
